// packages/bitcore-node/src/modules/ethereum/types.ts
export type IEthTransaction = ITransaction & {
  data: Buffer;
  gasLimit: number;
  gasPrice: number;
  nonce: number;
  to: string;
  from: string;
  internal: Array<ClassifiedTrace>;
  transactionIndex: number;
  abiType?: {
    type: string;
    name: string;
    params: Array<{ name: string; value: string; type: string }>;
  };
  error?: string;
  receipt?: {
    status: boolean;
    transactionHash: string;
    transactionIndex: number;
    blockHash: string;
    blockNumber: number;
    contractAddress?: string;
    cumulativeGasUsed: number;
    gasUsed: number;
    logs: Array<any>;
  };
};